fixed datatype

+++ src/http.cpp
@@ -277,7 +277,7 @@ size_t libhttppp::HttpResponse::printHeader(std::string &buffer){
 }
 
 
-void libhttppp::HttpResponse::send(netplus::con* curconnection,const char* data, unsigned long datalen){
+void libhttppp::HttpResponse::send(netplus::con* curconnection,const char* data, int datalen){
   if(datalen>=0){
         setContentLength(datalen);
   }
